backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
boto3==1.35.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
websockets==13.0
python-multipart==0.0.9
mangum==0.18.0
